# src/main/java/com/mendix/recipemanagement/models/Category.java
package com.mendix.recipemanagement.models;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class Category {
    private String categoryName;
}