# src/main/java/com/mendix/recipemanagement/models/Ingredient.java
package com.mendix.recipemanagement.models;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class Ingredient {
    private String item;
    private String quantity;
    private String unit;
}